backend/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
neo4j==5.9.0
faiss-cpu==1.7.4
openai==0.27.8
python-multipart==0.0.6
pyjwt==2.7.0
numpy==1.24.3
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==0.8.11